Added directions
Did not add autocomplete.
Most of this is Irfan's code, I cleaned it up a bit and added it on top
of Liam's last commit. It looks like it all works but it isn't limited
to Australia only even when I did add "Australia" onto the end of every
location string. If you could fix this it will be amazing.